TM Sprint-7 (0.7.2) source code delivered as is (neither rebuilt nor dry run tested yet at AtCor)
git-svn-id: https://sfs.atcorau.local/svn/SoftwareRepository/trunk/PN022@257 a8a9b14d-6b92-5b4b-9153-4cd90bb22c97

+++ Scor/dal/dal/DalDataBuffer.h
@@ -26,10 +26,10 @@ namespace AtCor{
 			public ref struct DalPwvDataStruct
 			{
 						public:
-							unsigned long tonometerData; /**< Tonometer data */
-							unsigned long cuffPulseData; /**< Cuff Pulse data */
-							unsigned long cuffPressure; /**< Cuff Pressure data */
-							unsigned long countdownTimer; /**< Countdown Timer interval */
+							unsigned short tonometerData; /**< Tonometer data */   
+							unsigned short cuffPulseData; /**< Cuff Pulse data */
+							unsigned short cuffPressure; /**< Cuff Pressure data */
+							unsigned short countdownTimer; /**< Countdown Timer interval */
 
 							/**
 							* Constructor for the structure. @n
